Modified json file, added delete confirmation and some minor fix

+++ src/view-employee-list/view-employee-list.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import {Employee} from '../Classes/EmployeeClass/employee';
 import {EmployeeServiceService} from '../Services/employee-service.service';
-import {MatSnackBar} from '@angular/material';
+import {MatDialog, MatSnackBar} from '@angular/material';
+import {DeleteConfirmationComponent} from '../delete-confirmation/delete-confirmation.component';
 
 
 @Component({
@@ -21,7 +22,7 @@ import {MatSnackBar} from '@angular/material';
 })
 export class ViewEmployeeListComponent implements OnInit {
 employeeList: Employee[];
-  constructor(private db: EmployeeServiceService, private snackBar: MatSnackBar) {
+  constructor(private db: EmployeeServiceService, private snackBar: MatSnackBar, private matDialog: MatDialog) {
   }
 
   ngOnInit() {
@@ -30,10 +31,17 @@ employeeList: Employee[];
     this.db.employeeArray.subscribe(array => this.employeeList = array);
   }
   deleteEmployee = (employee: Employee) => {
-    console.log('Going to delete: ' + employee.isDeleted);
-    this.employeeList = this.db.deleteEmployee(employee.employeeID, this.employeeList);
-    this.db.updateEmployeeList(this.employeeList);
-    this.snackBar.open('Employee deleted', '', {duration: 1000});
+    const ref = this.matDialog.open(DeleteConfirmationComponent);
+    ref.afterClosed().subscribe(isDelete => {
+      console.log('Delete: ' + typeof (isDelete));
+      if (isDelete === 'true') {
+        this.employeeList = this.db.deleteEmployee(employee.employeeID, this.employeeList);
+        this.db.updateEmployeeList(this.employeeList);
+        this.snackBar.open('Employee deleted', '', {duration: 1000});
+      } else {
+        console.log('Delete inside condition');
+        this.snackBar.open('Employee not removed', '', {duration: 1000});
+      }
+    });
   }
-
 }
